+++ tests/ops/api/v1/endpoints/test_connection_template_endpoints.py
@@ -48,7 +48,7 @@ class TestGetConnections:
         resp = api_client.get(url, headers=auth_header)
         data = resp.json()["items"]
         assert resp.status_code == 200
-        assert len(data) == 23
+        assert len(data) == 24
 
         assert {
             "identifier": ConnectionType.postgres.value,
@@ -157,7 +157,7 @@ class TestGetConnections:
         resp = api_client.get(url + "?system_type=database", headers=auth_header)
         assert resp.status_code == 200
         data = resp.json()["items"]
-        assert len(data) == 8
+        assert len(data) == 9
 
     def test_search_system_type_and_connection_type(
         self, api_client, generate_auth_header, url